prevention/bot/server/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
python-telegram-bot==20.0
requests
schedule
aiogram==3.0
apscheduler==3.8.1
python-dotenv==0.19.0